Add binary toggle for input
In the controller the previous version would
only allow the text box to display binary
numbers. This is a problem if the user wanted
to convert the number to decimal. The text
field wouldn't be able to accept the change
so I added a toggle to the controller to
so that decimal input could also be displayed.